+++ src/main/java/racingcar/model/Car.java
@@ -1,7 +1,5 @@
 package racingcar.model;
 
-import racingcar.model.wrapper.Name;
-import racingcar.model.wrapper.Position;
 import racingcar.strategy.RacingNumberGenerator;
 
 public class Car {
