+++ src/main/java/racingcar/model/Round.java
@@ -1,4 +1,4 @@
-package racingcar.wrapper;
+package racingcar.model;
 
 import racingcar.exception.BadRequestException;
 import racingcar.exception.ExceptionMessage;
